Move our local implementation of optional to public area.
optional was part of the public interface and not installing
it caused problems for projects that use kwiver. The include
file has been streamlined and moved to the public API.
